requirements.txt
Jinja2>=2.8
Pillow>=3.4.0
requests>=2.11.1
six
mss==6.1.0
numpy
opencv-contrib-python
facebook-wda>=1.3.3
pywinauto==0.6.3
